Check mode index not out-of-bounds before calling numba fast functions
Indeed, with the decorator @njit, the out-of-bound error is not raised
and can lead to random output when evaluating a solution.

Here we choose to set the solution as infeasible in that case, but one
could also prefer to raise an exception as it would do if we evaluate in
python counterparts with flage fast=True in do_recompute() or
run_sgs_partial().