Avoid parsing a data/ folder at module level
The goal:
We want to be able to import modules even though the data folder does
not exist.
This folder can be populated via utility functions available in
datasets module.
This folder will potentially not be located anymore in a subfolder of
source code (especially when installing the code not in development
mode)

The way to do so:
- avoid directly refering to data/ folder that should be in the library
  (for instance in .../site-packages/discrete-optimization/discrete_optimization/data)
- add argument to decide in which root directory all datasets should be
  found (consistent with utility functions used to download data in
  discrete-optimization.datasets)
- replace call to files_available or results_available module attributes
  to get_data_available() and get_results_available() functions