Set `objectives` argument as mandatory for GA, AlternatingGA and NSGA
In examples and tests, objectives are always explicitely given,
and `objectives` let to None was handled only in GA.evaluate_problem()
in the case of a single objective, but in a non-reproducible way.
(Based on the order of the keys of a dictionary)

This is an API change as we do not accept empty value for objectives
anymore.

We also fixed the way `objectives` is handled when it is a string instead
of a list of strings that is given.