Add a generic subclass for CPSolver wrapping minizinc
It allows to share the common solve() method between the various CP
solvers of the library.

As some solvers need to use a version of solve with a try..except around
the instance.solve(), we add a silent_solve_error attribute. When True
we add this try...except and raise a warning and return an empty
ResultStorage instead of raising an error.
Previously None was returned but it was not compliant with what expect
the solve() method.