Fix "Fetch data for tests" workflow step for windows
Since PR #48, minizinc is imported in
discrete_optimization.__init__.py to check the binary version, so even
when fetching tests data.

On windows we need to redefine the PATH properly so that minizinc
binary is found (while on mac and linux, having it defined at "Add
MiniZinc to PATH" step seems sufficient).