Update actions versions to avoid using node 12
Actions running on node 12 will not work anymore in summer 2023 and are
now raising warnings:
https://github.blog/changelog/2022-09-22-github-actions-all-actions-will-begin-running-on-node16-instead-of-node12/

We update all actions version to ones using node 16 instead of 12.

The only remaining ones is softprops/action-gh-release which has not
been updated since 2021. We should try to replace it by a more lively
one which made the update of node version.