Skip minizinc version check if proper env variable set
To ensure the library can work, we check in discrete_optimization.__init__.py
that minzinc binary is installed and configured with the correct version.

But in some cases, users want to use only a part of the library not
relying on minizinc. We allow this by skipping the check step
if the environment variable DO_SKIP_MZN_CHECK is set and evaluates as a
boolean to True.

This can be done for instance in shell with

	export DO_SKIP_MZN_CHECK=1