Change order of compute_lenght_***() arguments
We put "solution" as first argument so that partial functions
self.evaluate_function_2d() evaluate without named arguments.

Indeed up to now, to have the code work we needed

	self.evaluate_function_2d(solution=solution)

But mypy complained because it did not have the info that there were
named arguments and no positional argument

And now

	self.evaluate_function_2d(solution)

works either for mypy and at runtime.