Fix get_start_time and get_end_time
If we try to get time for a task not in the schedule, we rather raise a
KeyError instead of returning None. It avoids checking if the result is
None in places were it is expected to have returned an integer.

It is used in various place to be compared or added with other integers
for instance.