Add callback for optuna hyperparameters tuning
When using Optuna to tune hyperparameters of a solver,
one can use this callback to prune unpromising trials.

This callback reports every `report_nb_steps` an intermediate fit to optuna.

Avoid setting this parameter to 1 if using a solver with short steps,
as local search algorithms, which would excessively slow the run of a single trial.
Indeed, this pruning by optuna is rather designed
for epochs in deep learning trainers, which lasts quite a long time.

We add a test for this callback and thus add optuna to test
dependencies.

We do not add optuna to dependencies of discrete-optimization as the
library is functional without it. This is just a callback provided to
ease the usage of optuna to tune the hyperparameters.