Extract FirstSolutionStrategy and LocalSearchMetaheuristic enum from ortools
We can create dynamically the enumeration corresponding to
routing_enums_pb2.LocalSearchMetaheuristic and routing_enums_pb2.FirstSolutionStrategy.

We use them to feed init_model() so that the hyperparameter will be more
readable as a mere integer. We do not forget to cast them to the
corresponding integer when feeding the ortools routing model.

To be compatible with previous code, we allow the hyperparmeter to take
directly the integer value.