Fix test with TimerStopper callback
The callback is called only at each new solution discovery by the cp-sat
solver. So it may occur that the last solution is found before
the time limit set in the TimerStopper, but then the newt solution is not
found before the time limit set by parameters_cp.time_limit, so that the
callback TimerStopper is never called again.

This is the normal behaviour but prevent asserting the timerstopper
stopped the solve.

We fix this by
- reducing the time limit of TimerStopper
- freezing randomness to get reproducible tests