Fix ortools-cpsat-based solver when using several workers
When setting ParametersCP.nb_process > 1 (i.e.
ortools.sat.python.cp_model.CPSolver.num_workers >1), ortools use
multithreading and the handling of errors via pybind11 is not working
correctly. We got something like:
"terminate called after throwing an instance of 'pybind11::error_already_set'"
and the python aborts.

So instead of raising an error we store it in a dedicated attribute of
the do_solver and call callback.StopSearch(), and once outside ortools we
- either log it if it is an early stop due to a user callback returning
  True
- or re raise it if it is really an exception (for instance a
  TrialPruned raised by an OptunaPruningSingleFitCallback)