Implement callbacks for minizinc-based solvers
We use the instance.output_type as an entry point.
During the solve() call, we create dynamically a subclass of MinizincCPSolution
which
- wrapps an actual d-o solution
- has class attributes properly initialized for tracking
  - full ResultStorage
  - current number of step
  - solver instance
  - user-defined callbacks
This class will be called each time a new solution is found by minizinc,
and thus update the result storage and call upon user-defined callbacks.

If a stopping criteria is reached in callbacks, we raise a
SolveEarlyStop exception which will be caught in solve().

In addition to that a MinizincCPSolver must now implement `retrieve_solution()`
which converts a solution from minzinc into a d-o one.
It replaces retrieve_solutions() and means that solutions will be converted on the fly,
and directly added to the current result_storage.