Allow hyperparameter definition dependending on other hyperparameters
Sometimes, an hyperparameter has meaning only if another one has a
specific value (or is in a specific range of values).

Thus we define a new attribute `depends_on`
- if equals to None: no dependency is defined
- else should be (for now) a tuple `(hyperparameter_name, possible_values)`:
  the considered hyperparameter will be defined (and suggested by
  optuna) only if the hyperparameter whose name is `hyperparameter_name`
  have values falling in `possible_values`. `possible_values` must be
  of `Container` type so that we can test `value in possible_values`.
  For instance a list, a set or a pandas.Interval.

When suggesting hyperparameters, we suggest them in an order respecting
that depencies, by making use of
networkx.alogorithms.dag.topological_sort.

Note that SubBrickKwargsHyperparameter defining a
subbrick_hyperparameter attribute make also a kind of dependency
relationship that is taken into account in the suggestion order.

Thanks to that new mechanism, when suggesting hyperparameters with optuna,
we avoid having similar trials whose only differences are on irrelevant
hyperparameters. It makes optuna hyperparameter tuning clerarer and more efficient.