Add "name_in_kwargs" attribute to hyperparameters
If None, nothing changes.
Else the name_in_kwargs is used as the key for the hyperparameter in the
dictionary generated by `solver.suggest_hyperparameters_with_optuna()`.

This is useful when several hyperparameters are defined with different
characteristics, are depending on others so that only one is suggested
at a time among them, and are all expected to have the same name in
solver.__init__. For instance if they are the parameters of a submethod
which is not controlled by the library and happen to share names.