Add possibility to suggest bounds of a float hyperparameter
Thanks to the new attributes (and corresponding `suggest_with_optuna()`
arguments) `suggest_low` and `suggest_high`, we enforce the
possibility of suggesting lower or upper bound.
Indeed, by default, optuna will suggest inside the range (low, high) but
never reach exactly a bound (except if a step was provided).

To achieve this, a derived categorical hyperparameter is suggested under
the hood to decide whether suggesting the lower bound, upper bound, or a
float in between. The "generated" hyperparameter will be named as the
original hyperparameter, suffixed with ".suggest_bound" and will take as
values
- "low": lower bound suggested
- "high": upper bound suggested
- "": usual behaviour