Fix lns solve() and unify contraint handlers api
Do not forget to remove new constraints at the end of each lns
iteration.
For minizinc lns, it is not doing anything are this is manage by the
child instance creation but for mip lns, this is relevant.

We implement remove_constraints_from_previous_iteration() once for all
in base classes for
- minizinc based constraint handlers (by doing nothing).
- gurobi based constraint handlers
- pymip based constraint handlers

We also unify their api so that the return value for adding_constraint_from_results_store()
(and type of previous_constraints in remove_constraints_from_previous_iteration)
is always an iterable of constraints. (no more dictionaries)

Add a common __init__ for PymipMilpSolver to ensure that each pymip solvers have
a milp_solver_name attribute needed by the constraint handler remove***() method.