Fix use of time_limit_iter0 in LNS_CP
Instead of passing iteration (which was wrongly always pass as 0) to
solve_with_subsolver(), we pass a boolean `no_previous_solution` which
will be true only if we skipped the initial_solution_provider and we are
at the first lns iteration. In that case we will use time_limit_iter0
instead of time_limit to start from a "good" solution.

We also
- rename skip_first_iteration into skip_initial_solution_provider
- skip the initial solution provider also if
  self._initial_solution_provider is None

NB: we keep the boolean for optuna hyperparameters suggestion