Add a unit when reporting time with optuna callback
Starting from optuna 4.0, report steps are automatically converted to
integers (as they are meant to be) so that our "trick" of passing the
time instead may result in intermediate value being overriden if steps
are within the same second.
We thus introduce a unit factor (if 1., nothing change, if 0.001 it
converts time to milliseconds) to be able to have a finer granularity.
This is useful to make the corresponding unit test still work without
slowing it.