Split sequential metasolver hyperparameter into first and next subsolvers
We split former hyperparameter "list_subbricks" into
- subsolver_0 (the first one)
- next_subsolvers (the others)

This avoids issues when specifying subsolvers non-warmstartable,
as they will be valid only for the first subsolver.

The complete list of subsolvers is then reconstructed in __init__().

NB: this concerns only the auto generation by optuna of the list of
subsolvers.

In order to have the first subsolver generated by "next_subsolvers"
hyperparameter be named "subsolver_1", and thus not overriding
"subsolver_0", we also add an option in ListHyperparameters for deciding
the start of numbering of the hyperparameters generated.