Add status to lp solvers (and default one to all solvers)
Gurobi and mathopt give information about solver status. As for minizinc
and cpsat, we make it available on the wrapper level.

As we use the same enumeration `StatusSolver`, we move it to `solver_do`
module. And thus add for all solvers a `status_solver` attribute, by
default equal to unknown and which is used by `is_optimal()` method.

We remove `get_status_solver()` method which was only returning
`self.status_solver`.