New mixin BoundsProviderMixin to have generic gap based early stoppers
We create a new abstract mixin for solvers exposing
- current internal objective value
- current internal objective (dual) bound

These solver can thus compute a (relative or absolute) gap at each
iteration in a user callback. So we implement a generic stopper based on
gap (following the existing one for CP-SAT) and also a stats retriever
callback that stores also these objective values/bounds.

The mixin api is implemented for
- cpsat based solvers
- mathopt based solvers
- gurobi based solvers