Fix interactive example on unsatisfiability with fine constraints
Because of the overriding of solver.get_soft_constraints(), removing
constraints from the model was not sufficient to be taken into account
by solver.get_soft_constraints(). So we explicitely track the soft
constraints passed to solver.explain_unsat_fine() to avoid looping
infinitely on the same choice of constraints.