Use ruff linter+formatter instead of black, isort, and pycln
Ruff has been added to the pre-commit hooks to replace black and isort (both for python files and notebooks).

The changes have been made by running pre-commit on all files

	pre-commit run -a

ruff-format is very close to black in terms of formatting result and much faster.
The ruff linter offers lots of possibilities (with same speed enhancement as the formatter) but is only used for now to replace isort, and pycln in notebooks.