Make explicit the need for `tasks_list` and `unary_resources_list`
The child classes of `TasksProblem` and `AllocationProblem` need to implement
the definition of tasks_list and unary_resources_list. In python the
easiest way to declare them as abstract is to define them as (abstract)
properties. See for instance https://discuss.python.org/t/abstract-variables-in-abc/69071/6

It means that child classes need also to declare them as property. This
was already the case except for rcpsp, rcpsp_multiskill and vrptw that
set them as attributes. We make the necessary modification in their
`__init__()`.

Any "good IDE" should now tell someone wanting to implement these
interfaces that they need to implement `tasks_list` and
`unary_resources_list`. And not doing so will raise an error at runtime.